refactor: moved joystick options to separate files
