+++ Assets/PlayroomKit/modules/Options/ButtonOptions.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Playroom
+{
+    [Serializable]
+    public class ButtonOptions
+    {
+        public string id = null;
+        public string label = "";
+        public string icon = null;
+    }
+}
